affiliation directoy refactored
