+++ api-js/src/affiliation/objects/__tests__/affiliation-connection.test.js
@@ -1,5 +1,5 @@
 import { GraphQLInt } from 'graphql'
-import { affiliationConnection } from '../affiliation'
+import { affiliationConnection } from '../affiliation-connection'
 
 describe('given the affiliation connection object', () => {
   describe('testing its field definitions', () => {
