+++ api-js/src/affiliation/objects/index.js
@@ -1,1 +1,2 @@
 export * from './affiliation'
+export * from './affiliation-connection'
